+++ src/main/java/com/ems/service/EmployeeService.java
@@ -9,8 +9,8 @@ public interface EmployeeService {
 
     public List<Employee> getAllEmployees();
     public Long saveEmployee(Employee employee);
-    public Employee getEmployeById(Long id);
-    public void deleteById(Long id);
-    public  boolean isExist(Long id);
+    public Employee getEmployeById(Integer id);
+    public void deleteById(Integer id);
+    public  boolean isExist(Integer id);
     public List<Employee> getByFirstName(String firstName);
 }
